requirements.txt
pydantic~=1.10.4
fastapi~=0.89.1
starlette~=0.22.0
uvicorn~=0.20.0
python-dotenv~=0.21.1